refactor: move stores inside utils
