+++ src/routes/posts/index.svelte
@@ -20,8 +20,8 @@
 	import LayoutPicker from '../../pages/LayoutPicker.svelte';
 	import PostCard from '../../components/PostCard.svelte';
 
-	import { pSlice as store } from '../../stores';
 	import { sieve, filter } from '../../utils/search';
+	import { pSlice as store } from '../../utils/stores';
 	let query, filtered, sieved;
 	let filters = { categories: [], tags: [], sort_by: 'updated' };
